[JBAOP-256] The previous solution was not working correctly on all possible scenarios.
git-svn-id: https://svn.jboss.org/repos/jbossas/projects/aop/trunk@68854 84be2c1e-ba19-0410-b317-a758671a6fc1

+++ aop/src/main/org/jboss/aop/standalone/Compiler.java
@@ -263,7 +263,12 @@ public class Compiler
       {
          String fileName = file.getName().substring(0, index) + ".class";
          File superClassFile = new File(fileName);
-         if (!loadFile(superClassFile))
+         // checking last modified date is not enough
+         // because these values can differ in some miliseconds sometimes
+         if (superClassFile.lastModified() > file.lastModified()
+         // so, check this class has been recompiled and, hence,
+         // is not Advised anymore
+               && !loadFile(superClassFile))
          {
             file.delete();
             return;
